requirements.txt
yfinance
pandas
pandas_ta
google-generativeai
duckduckgo-search
requests
python-dotenv
